More Buttons!
Added two new buttons, and the board starts off with no pieces. When you click the set up button it puts all the pieces in the start position, and enables the Reset button (does the same thing as before), and the Clear Board which returns the board to it's initial state.