Tabs, Tabs, and More Tabs!
Created an element to the side of the board where one can click on the tabs to get different options. I moved the chess piece image on the sides into the html, and just have both colors instead of a button to change their color. Minor refactoring throughout. Added error image if image is not loaded, so as to not create invisible pieces on the board. And fixed dragging issue where you could drag anything onto the board, and if would create and invisible piece.