Refactoring, and Bug Fix
Added new tab: Customize that doesn't do anything yet. Changed all the hide and show methods to one, and one event that all the Tab buttons are subcribed to. Fixed a bug where the set-up board was removing the picture of a placed chess piece, but not clearing it from the tiles, and changed sizing of tab controls.