Refactoring!
Refactored out the event, and movment logic for a chessboard, and removed the refrence of chesspieces from the chess board. Also, removed the reference that peice and tile had to each other respectively. Moved all this to a controller, that controlls the piece and board seperately and handles all the events. Added point object to simplify code: instead of passing left, and top, you can pass one object leftTop that has an x, and a y.